Clean up debug sign conversions in deflate check_match calls.
Cast match_len/prev_length to int at check_match call sites to
avoid implicit unsigned-to-signed conversion warnings.

Upstream: https://github.com/madler/zlib/commit/f4f34491

+++ deflate_medium.c
@@ -37,7 +37,7 @@ static int emit_match(deflate_state *s, struct match match) {
         return bflush;
     }
 
-    check_match(s, match.strstart, match.match_start, match_len);
+    check_match(s, match.strstart, match.match_start, (int)match_len);
 
     bflush += zng_tr_tally_dist(s, match.strstart - match.match_start, match_len - STD_MIN_MATCH);
     return bflush;
